chore: move GradientTransition into its own directory; docs: update documentation of changed components
