+++ src/lib/Components/Layout/ShowcaseSection.svelte
@@ -4,6 +4,8 @@
     These sorts of sections are comprised of a title placed on the top-left corner of the screen, and the content goes beneath it, whichever shape it may take.
     It wraps around the PageSection component, assuring consistent placement and size of the Title and remaining content.
 	The content is injected using svelte's <slot> element
+		
+		@param title - string - Title of this ShowcaseSection
 -->
 <script lang="ts">
 	import PageSection from "./PageSection.svelte";
