+++ src/lib/Components/Project/Project.svelte
@@ -2,7 +2,6 @@
 @component
 	This component renders a project element. It consists of a (actually multiple) gradient background(s), a background image, some text, and some other text that is only displayed when hovered.
 
-		@param gridID - number - number of the item on the projects grid
 		@param projectData - ProjectData - holds the data which will be displayed on this element
 -->
 <script lang="ts">
@@ -38,7 +37,7 @@
 					title="My Contributions"
 					titleSize="3xl"
 					paragraphs={projectData.contributions}
-					descriptionBullets={true}
+					paragraphBullets={true}
 				/>
 			</div>
 		{/if}
